remove abstrat
