+++ monai/apps/detection/metrics/__init__.py
@@ -8,5 +8,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from monai.apps.detection.metrics.abstract import AbstractEvaluator, AbstractMetric, DetectionMetric
